Refined connectivity analysis
When using the analysis-interval arg, the connectivity analysis will run mid-simulation only until the graph is strongly connected, and the plot will be displayed only until it is weakly connected.
At the end of a simulation, we also print the first time_step in which the graph became strongly/weakly connected.